backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk[database]==0.3.0",
    "yfinance==0.2.56",
    "psutil==5.9.5",
    "litellm==1.66.3",
    "google-generativeai==0.8.5",
    "python-dotenv==1.1.0",
    "Deprecated>=1.2.14",
    "vercel-ai",
    "fastapi[standard]",
    "uvicorn[standard]",
    "groq",
    "sqlalchemy",
    "psycopg2-binary",
    "alembic",
    "passlib",
    "bcrypt",
    "python-jose[cryptography]",
    "pydantic-settings",
    "jira>=3.10.5",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "httpx>=0.24.0",
    "respx>=0.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "agents*"]

